+++ packages/apollo-client/src/scheduler/__tests__/scheduler.ts
@@ -220,41 +220,6 @@ describe('QueryScheduler', () => {
     });
   });
 
-  it('should handle graphql errors on polling queries correctly', done => {
-    const query = gql`
-      query {
-        author {
-          firstName
-          lastName
-        }
-      }
-    `;
-    const errors = [new Error('oh no something went wrong')];
-    const queryOptions = {
-      query,
-      pollInterval: 80,
-    };
-    const link = mockSingleLink({
-      request: queryOptions,
-      result: { errors },
-    });
-    const queryManager = new QueryManager({
-      link,
-      store: new DataStore(new InMemoryCache()),
-    });
-    const scheduler = new QueryScheduler({
-      queryManager,
-    });
-    let observableQuery = scheduler.registerPollingQuery(queryOptions);
-    const subscription = observableQuery.subscribe({
-      error(errorVal) {
-        subscription.unsubscribe();
-        expect(errorVal).toBeDefined();
-        done();
-      },
-    });
-  });
-
   it('should not fire another query if one with the same id is in flight', done => {
     const query = gql`
       query B {