+++ packages/apollo-client/CHANGELOG.md
@@ -2,6 +2,10 @@
 
 ### vNext
 
+### 2.0.0-beta.7
+- Move graphql to peerDependency
+- Ensure network errors don't cause unhandled rejections on polled queries
+
 ### 2.0.0-beta.6
 - Support conditional refetches for mutations
 - Ensure network errors don't cause unhandled rejections on cache-and-network policies